fix(mcp-analytics): dash the dashboard's in-progress bucket
activityIncompleteTail was computed and covered by its own tests, and
ActivityChart draws the partial stroke from it, but the scene never passed the
prop. On the default last-7-days view the final bucket is a few hours of data
plotted against full days, so it drops and reads as a collapse in tool calls
rather than a period that hasn't finished.

Making the prop required rather than optional is what keeps it wired: an
omitted optional prop is exactly how this shipped.

lastBucketIsInProgress moves to timeBuckets, next to the window and bucket-key
primitives it belongs with, ahead of the other MCP analytics surfaces needing
it.
