+++ products/mcp_analytics/frontend/MCPAnalyticsDashboardOverview.tsx
@@ -30,6 +30,7 @@ export function MCPAnalyticsDashboardOverview(): JSX.Element {
         harnessRowsLoading,
         dailyActivity,
         activityRowsLoading,
+        activityIncompleteTail,
         toolDailySeries,
         toolDailyRowsLoading,
         toolRows,
@@ -102,6 +103,7 @@ export function MCPAnalyticsDashboardOverview(): JSX.Element {
                                 theme={theme}
                                 timezone={timezone}
                                 interval={interval}
+                                incompleteTail={activityIncompleteTail}
                             />
                         </div>
                         <HarnessDonut rows={harnessRows} loading={harnessRowsLoading} theme={theme} />
