+++ products/mcp_analytics/frontend/dashboard/DashboardCards.stories.tsx
@@ -28,6 +28,14 @@ const DAILY_ACTIVITY: DailyActivity = {
     errors: [120, 140, 160, 170, 180, 176, 168],
 }
 
+// The shape the in-progress stories exist for: the last bucket is a few hours into the day, so its
+// counts sit far below its neighbours. Dashing it is what stops that reading as a collapse.
+const DAILY_ACTIVITY_PARTIAL_TAIL: DailyActivity = {
+    labels: DAYS,
+    successes: [4180, 4360, 4560, 4430, 4720, 4920, 1680],
+    errors: [120, 140, 160, 170, 180, 176, 54],
+}
+
 const TOOL_DAILY: ToolDailySeries = {
     labels: DAYS,
     tools: [
@@ -154,7 +162,27 @@ export const KeyMetrics: Story = {
 
 export const DailyCallsAndErrors: Story = {
     render: withTheme((theme) => (
-        <ActivityChart daily={DAILY_ACTIVITY} loading={false} theme={theme} timezone="UTC" interval="day" />
+        <ActivityChart
+            daily={DAILY_ACTIVITY}
+            loading={false}
+            theme={theme}
+            timezone="UTC"
+            interval="day"
+            incompleteTail={false}
+        />
+    )),
+}
+
+export const DailyCallsAndErrorsInProgressBucket: Story = {
+    render: withTheme((theme) => (
+        <ActivityChart
+            daily={DAILY_ACTIVITY_PARTIAL_TAIL}
+            loading={false}
+            theme={theme}
+            timezone="UTC"
+            interval="day"
+            incompleteTail
+        />
     )),
 }
